все идет медленно

+++ utils/db_api/schemas/referral_program.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, BigInteger, String, sql
+from sqlalchemy import Integer, Column, BigInteger, String, sql, Float
 
 from utils.db_api.db_gino import TimedBaseModel
 
@@ -7,6 +7,6 @@ class Referral(TimedBaseModel):
     __tablename__ = 'referrals'
     user_id = Column(BigInteger, primary_key=True)
     user_name = Column(String(100))
-    balance = Column(Integer(), default=0)
+    balance = Column(Float(precision=6), default=0.0)
 
     query: sql.Select
